+++ service/flex/v1/plugin/api_op_fetch.go
@@ -25,12 +25,14 @@ type FetchPluginResponse struct {
 // Fetch retrieves a plugin resource
 // See https://www.twilio.com/docs/flex/developer/plugins/api/plugin#fetch-a-plugin-resource for more details
 // Context is defaulted to Background. See https://golang.org/pkg/context/#Background for more information
+// This resource is currently in beta and subject to change. Please use with caution
 func (c Client) Fetch() (*FetchPluginResponse, error) {
 	return c.FetchWithContext(context.Background())
 }
 
 // FetchWithContext retrieves a plugin resource
 // See https://www.twilio.com/docs/flex/developer/plugins/api/plugin#fetch-a-plugin-resource for more details
+// This resource is currently in beta and subject to change. Please use with caution
 func (c Client) FetchWithContext(context context.Context) (*FetchPluginResponse, error) {
 	op := client.Operation{
 		Method: http.MethodGet,