+++ service/flex/v1/plugin/version/api_op_fetch.go
@@ -26,12 +26,14 @@ type FetchVersionResponse struct {
 // Fetch retrieves a plugin version resource
 // See https://www.twilio.com/docs/flex/developer/plugins/api/plugin-version#fetch-a-pluginversion-resource for more details
 // Context is defaulted to Background. See https://golang.org/pkg/context/#Background for more information
+// This resource is currently in beta and subject to change. Please use with caution
 func (c Client) Fetch() (*FetchVersionResponse, error) {
 	return c.FetchWithContext(context.Background())
 }
 
 // FetchWithContext retrieves a plugin version resource
 // See https://www.twilio.com/docs/flex/developer/plugins/api/plugin-version#fetch-a-pluginversion-resource for more details
+// This resource is currently in beta and subject to change. Please use with caution
 func (c Client) FetchWithContext(context context.Context) (*FetchVersionResponse, error) {
 	op := client.Operation{
 		Method: http.MethodGet,