+++ service/flex/v1/plugin/versions/api_op_create.go
@@ -34,12 +34,14 @@ type CreateVersionResponse struct {
 // Create creates a new plugin version resource
 // See https://www.twilio.com/docs/flex/developer/plugins/api/plugin-version#create-a-pluginversion-resource for more details
 // Context is defaulted to Background. See https://golang.org/pkg/context/#Background for more information
+// This resource is currently in beta and subject to change. Please use with caution
 func (c Client) Create(input *CreateVersionInput) (*CreateVersionResponse, error) {
 	return c.CreateWithContext(context.Background(), input)
 }
 
 // CreateWithContext creates a new plugin version resource
 // See https://www.twilio.com/docs/flex/developer/plugins/api/plugin-version#create-a-pluginversion-resource for more details
+// This resource is currently in beta and subject to change. Please use with caution
 func (c Client) CreateWithContext(context context.Context, input *CreateVersionInput) (*CreateVersionResponse, error) {
 	op := client.Operation{
 		Method:      http.MethodPost,