+++ service/flex/v1/plugin/versions/api_op_page.go
@@ -51,12 +51,14 @@ type VersionsPageResponse struct {
 // Page retrieves a page of plugin versions
 // See https://www.twilio.com/docs/flex/developer/plugins/api/plugin-version#read-multiple-pluginversion-resources for more details
 // Context is defaulted to Background. See https://golang.org/pkg/context/#Background for more information
+// This resource is currently in beta and subject to change. Please use with caution
 func (c Client) Page(options *VersionsPageOptions) (*VersionsPageResponse, error) {
 	return c.PageWithContext(context.Background(), options)
 }
 
 // PageWithContext retrieves a page of plugin versions
 // See https://www.twilio.com/docs/flex/developer/plugins/api/plugin-version#read-multiple-pluginversion-resources for more details
+// This resource is currently in beta and subject to change. Please use with caution
 func (c Client) PageWithContext(context context.Context, options *VersionsPageOptions) (*VersionsPageResponse, error) {
 	op := client.Operation{
 		Method: http.MethodGet,