+++ service/flex/v1/plugin_configuration/api_op_client.go
@@ -9,6 +9,7 @@ import (
 
 // Client for managing a specific plugin configuration resource
 // See https://www.twilio.com/docs/flex/developer/plugins/api/plugin-configuration for more details
+// This client is currently in beta and subject to change. Please use with caution
 type Client struct {
 	client *client.Client