+++ service/flex/v1/plugin_configuration/api_op_fetch.go
@@ -23,12 +23,14 @@ type FetchConfigurationResponse struct {
 // Fetch retrieves a plugin configuration resource
 // See https://www.twilio.com/docs/flex/developer/plugins/api/plugin-configuration#fetch-a-pluginconfiguration-resource for more details
 // Context is defaulted to Background. See https://golang.org/pkg/context/#Background for more information
+// This resource is currently in beta and subject to change. Please use with caution
 func (c Client) Fetch() (*FetchConfigurationResponse, error) {
 	return c.FetchWithContext(context.Background())
 }
 
 // FetchWithContext retrieves a plugin configuration resource
 // See https://www.twilio.com/docs/flex/developer/plugins/api/plugin-configuration#fetch-a-pluginconfiguration-resource for more details
+// This resource is currently in beta and subject to change. Please use with caution
 func (c Client) FetchWithContext(context context.Context) (*FetchConfigurationResponse, error) {
 	op := client.Operation{
 		Method: http.MethodGet,