+++ service/flex/v1/plugin_configuration/plugins/api_op_client.go
@@ -4,6 +4,7 @@ package plugins
 import "github.com/RJPearson94/twilio-sdk-go/client"
 
 // Client for managing plugin resources
+// This client is currently in beta and subject to change. Please use with caution
 type Client struct {
 	client *client.Client