+++ service/flex/v1/plugin_releases/api_op_page.go
@@ -46,12 +46,14 @@ type ReleasesPageResponse struct {
 // Page retrieves a page of plugin release resources
 // See https://www.twilio.com/docs/flex/developer/plugins/api/release#read-multiple-pluginrelease-resources for more details
 // Context is defaulted to Background. See https://golang.org/pkg/context/#Background for more information
+// This resource is currently in beta and subject to change. Please use with caution
 func (c Client) Page(options *ReleasesPageOptions) (*ReleasesPageResponse, error) {
 	return c.PageWithContext(context.Background(), options)
 }
 
 // PageWithContext retrieves a page of plugin release resources
 // See https://www.twilio.com/docs/flex/developer/plugins/api/release#read-multiple-pluginrelease-resources for more details
+// This resource is currently in beta and subject to change. Please use with caution
 func (c Client) PageWithContext(context context.Context, options *ReleasesPageOptions) (*ReleasesPageResponse, error) {
 	op := client.Operation{
 		Method:      http.MethodGet,