+++ service/flex/v1/plugins/api_op_page.go
@@ -50,12 +50,14 @@ type PluginsPageResponse struct {
 // Page retrieves a page of plugins
 // See https://www.twilio.com/docs/flex/developer/plugins/api/plugin#read-multiple-plugin-resources for more details
 // Context is defaulted to Background. See https://golang.org/pkg/context/#Background for more information
+// This resource is currently in beta and subject to change. Please use with caution
 func (c Client) Page(options *PluginsPageOptions) (*PluginsPageResponse, error) {
 	return c.PageWithContext(context.Background(), options)
 }
 
 // PageWithContext retrieves a page of plugins
 // See https://www.twilio.com/docs/flex/developer/plugins/api/plugin#read-multiple-plugin-resources for more details
+// This resource is currently in beta and subject to change. Please use with caution
 func (c Client) PageWithContext(context context.Context, options *PluginsPageOptions) (*PluginsPageResponse, error) {
 	op := client.Operation{
 		Method:      http.MethodGet,